fix: use `SidePanelParentRoom` as a data wrapper to provide the subscription instead of throwing error
