+++ apps/meteor/client/views/navigation/sidepanel/SidepanelItem/SidePanelParentRoom/SidePanelParentRoomWithData.tsx
@@ -1,0 +1,15 @@
+import { useUserSubscription } from '@rocket.chat/ui-contexts';
+
+import SidePanelParentRoom from './SidePanelParentRoom';
+
+const SidePanelParentRoomWithData = ({ prid }: { prid: string }) => {
+	const subscription = useUserSubscription(prid);
+
+	if (!subscription) {
+		return null;
+	}
+
+	return <SidePanelParentRoom subscription={subscription} />;
+};
+
+export default SidePanelParentRoomWithData;
