+++ apps/meteor/client/views/navigation/sidepanel/SidepanelItem/SidePanelParentRoom/index.ts
@@ -1,0 +1,1 @@
+export { default } from './SidePanelParentRoomWithData';
